Renamed password input field
